+++ src/Loader.php
@@ -8,6 +8,7 @@ class Loader
 {
     public function __construct()
     {
+        /** @noinspection PhpIncludeInspection */
         require_once 'recipe/common.php';
         new Load([
                 ['path' => 'vendor/sourcebroker/deployer-extended/deployer'],