+++ deployer/magento/task/magento:setup:static-content:deploy:extended.php
@@ -2,14 +2,6 @@
 
 namespace Deployer;
 
-task('magento:deploy:git_checkout', function () {
-    $activePath = get('deploy_path') . '/' . (test('[ -L {{deploy_path}}/release ]') ? 'release' : 'current');
-    foreach (get('magento_git_checkout_items', []) as $checkoutItem) {
-        run('cd ' . $activePath . ' && git checkout ' . escapeshellarg($checkoutItem));
-    }
-})->desc('Checkout for files/folders on git repo.');
-
-
 task('magento:setup:static-content:deploy:extended', function () {
     $areas = ['frontend', 'adminhtml'];
     foreach ($areas as $area) {
@@ -18,10 +10,10 @@ task('magento:setup:static-content:deploy:extended', function () {
         }, get('static_content_deploy_' . $area . '_themes', [])));
 
         $staticContentDeployLanguagesCmd = implode(' ', array_map(function ($language) {
-            return '--language ' . $language;
+            return ' ' . $language;
         }, get('static_content_deploy_' . $area . '_languages', [])));
 
-        run('cd {{release_path}} && {{bin/php}} bin/magento setup:static-content:deploy ' .
+        run('cd {{release_path}} && {{bin/php}} bin/magento setup:static-content:deploy -a ' . $area . ' '.
             $staticContentDeployLanguagesCmd . ' ' . $staticContentDeployThemesCmd);
     }
 })->desc('Static-content deploy for declared themes and languages.');