+++ deployer/deploy/task/deploy.php
@@ -9,48 +9,60 @@ task('deploy', [
     // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-composer-install
     'deploy:check_composer_install',
 
-    // Standard deployer deploy:prepare
+    // Standard Deployer "deploy:prepare" command.
     'deploy:prepare',
 
-    // Standard deployer deploy:lock
+    // Standard Deployer "deploy:lock" command.
     'deploy:lock',
 
-    // Standard deployer deploy:release
+    // Standard Deployer "deploy:release" command.
     'deploy:release',
 
-    // Standard deployer deploy:update_code
+    // Standard Deployer "deploy:update_code" command.
     'deploy:update_code',
 
-    // Standard deployer deploy:shared
+    // Standard Deployer "deploy:shared" command.
     'deploy:shared',
 
-    // Standard deployer deploy:writable
+    // Standard Deployer "deploy:writable" command.
     'deploy:writable',
 
-    // Standard deployer deploy:vendors
+    // Standard Deployer "deploy:vendors" command.
     'deploy:vendors',
 
-    // Git checkout for files overwritten while composer install for magento/magento2-base
-    'magento:deploy:git_checkout',
-
-    // Standard deployer deploy:clear_paths
+    // Standard Deployer "deploy:clear_paths" command.
     'deploy:clear_paths',
 
     // Clear php cli cache.
     // Read more on https://github.com/sourcebroker/deployer-extended#php-clear-cache-cli
     'php:clear_cache_cli',
 
-    // Start buffering http requests. No frontend access possible from now.
-    // Read more on https://github.com/sourcebroker/deployer-extended#buffer-start
-    'buffer:start',
+    // Git checkout for files overwritten while composer install for magento/magento2-base
+    'magento:deploy:git_checkout',
 
-    // Call "magento setup:upgrade"
-    'magento:setup:upgrade',
+    // Standard Magento "setup:di:compile" command.
+    'magento:setup:di:compile',
 
     // Call "magento setup:static-content:deploy" with parameters for themes and languages
     'magento:setup:static-content:deploy:extended',
 
-    // Standard deployer symlink (symlink release/x/ to current/)
+    // Start buffering http requests. No frontend access possible from now.
+    // Read more on https://github.com/sourcebroker/deployer-extended#buffer-start
+    'buffer:start',
+
+    // Standard Magento "setup:db:schema:upgrade" command.
+    'magento:setup:db:schema:upgrade',
+
+    // Standard Magento "setup:db:data:upgrade" command.
+    'magento:setup:db:data:upgrade',
+
+    // Standard Magento "app:config:import" command.
+    'magento:app:config:import',
+
+    // Standard Magento "cache:flush" command.
+    'magento:cache:flush',
+
+    // Standard Deployer "deploy:symlink" (symlink release/x/ to current/)
     'deploy:symlink',
 
     // Clear frontend http cache.
@@ -61,9 +73,9 @@ task('deploy', [
     // Read more on https://github.com/sourcebroker/deployer-extended#buffer-stop
     'buffer:stop',
 
-    // Standard deployer deploy:unlock
+    // Standard Deployer "deploy:unlock" command.
     'deploy:unlock',
 
-    // Standard deployer cleanup.
+    // Standard Deployer "cleanup" command.
     'cleanup',
-])->desc('Deploy your Magento2');
+])->desc('Deploy your Magento 2.2');