+++ deployer/deploy/task/deploy.php
@@ -3,11 +3,21 @@
 namespace Deployer;
 
 task('deploy', [
+
+    // Standard deployer task.
+    'deploy:info',
+
     // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-lock
     'deploy:check_lock',
 
     // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-composer-install
     'deploy:check_composer_install',
+
+    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-branch-local
+    'deploy:check_branch_local',
+
+    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-check-branch
+    'deploy:check_branch',
 
     // Standard Deployer "deploy:prepare" command.
     'deploy:prepare',
@@ -36,9 +46,13 @@ task('deploy', [
     // Standard Deployer "deploy:clear_paths" command.
     'deploy:clear_paths',
 
+    // Create database backup, compress and copy to database store.
+    // Read more on https://github.com/sourcebroker/deployer-extended-database#db-backup
+    'db:backup',
+
     // Clear php cli cache.
-    // Read more on https://github.com/sourcebroker/deployer-extended#php-clear-cache-cli
-    'php:clear_cache_cli',
+    // Read more on https://github.com/sourcebroker/deployer-extended#cache-clear-php-cli
+    'cache:clear_php_cli',
 
     // Standard Magento "setup:di:compile" command.
     'magento:setup:di:compile',
@@ -65,17 +79,28 @@ task('deploy', [
     // Standard Deployer "deploy:symlink" (symlink release/x/ to current/)
     'deploy:symlink',
 
+    // Clear php cli cache.
+    // Read more on https://github.com/sourcebroker/deployer-extended#cache-clear-php-cli
+    'cache:clear_php_cli',
+
     // Clear frontend http cache.
-    // Read more on https://github.com/sourcebroker/deployer-extended#php-clear-cache-http
-    'php:clear_cache_http',
+    // Read more on https://github.com/sourcebroker/deployer-extended#cache-clear-php-http
+    'cache:clear_php_http',
 
     // Frontend access possible again from now
     // Read more on https://github.com/sourcebroker/deployer-extended#buffer-stop
     'buffer:stop',
 
-    // Standard Deployer "deploy:unlock" command.
+    // Standard deployer task.
     'deploy:unlock',
 
-    // Standard Deployer "cleanup" command.
+    // Standard deployer task.
     'cleanup',
+
+    // Read more on https://github.com/sourcebroker/deployer-extended#deploy-extend-log
+    'deploy:extend_log',
+
+    // Standard deployer task.
+    'success',
+
 ])->desc('Deploy your Magento 2.2');