+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Changelog
 
 a) [TASK] Use deployer-loader to load recipes.
 b) [TASK] Add PhpIncludeInspection to prevent warnings in IDE.
+c) [TASK][!!!] Update packages sourcebroker/deployer-extended, sourcebroker/deployer-extended-media,
+   sourcebroker/deployer-extended-database, sourcebroker/deployer-loader. Add package sourcebroker/deployer-loader.
 
 0.7.1
 ~~~~~