+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ a) [TASK] Use deployer-loader to load recipes.
 b) [TASK] Add PhpIncludeInspection to prevent warnings in IDE.
 c) [TASK][!!!] Update packages sourcebroker/deployer-extended, sourcebroker/deployer-extended-media,
    sourcebroker/deployer-extended-database, sourcebroker/deployer-loader. Add package sourcebroker/deployer-loader.
+d) [TASK] Add .gitignore with vendors/ and composer.lock.
 
 0.7.1
 ~~~~~