+++ CHANGELOG.rst
@@ -19,7 +19,7 @@ j) [TASK] Extend "shared_files" and "writable_dirs".
 k) [TASK][!!!] Change deploy tasks order / add new tasks for Magento 2.2.
 l) [TASK][!!!] Change name of "magento:deploy:git_checkout" to "magento:deploy:git-checkout" to follow magento convention.
 m) [TASK] Make "magento:deploy:git-checkout" to be run before "deploy:clear_paths".
-
+n) [TASK] Replace occurance 'pub' in deployer/set.php with {{web_path}}
 
 0.7.1
 ~~~~~